Create a new component specific to the script dialog's content.
